% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RationalFunctions.R
\name{as.character.Rational}
\alias{as.character.Rational}
\title{Rational to string}
\usage{
\method{as.character}{Rational}(x)
}
\arguments{
\item{x}{A `Rational` object}
}
\value{
Character version of `x`
}
\description{
Rational to string
}
