% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gcd}
\alias{gcd}
\alias{lcm}
\title{GCD and LCM functions}
\usage{
gcd(x, y)

lcm(x, y)
}
\arguments{
\item{x}{A numeric object that can be coerced to integer}

\item{y}{A numeric object that can be coerced to integer}
}
\value{
The GCD or LCM
}
\description{
GCD and LCM functions
}
