% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mydat}
\alias{mydat}
\title{Sample data set
Does some stuff}
\format{
An object of class \code{data.frame} with 5 rows and 2 columns.
}
\usage{
mydat
}
\description{
Sample data set
Does some stuff
}
\keyword{datasets}
