% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RationalFunctions.R
\name{quotient}
\alias{quotient}
\title{Convert a `Rational` into a decimal}
\usage{
quotient(object, digits = NULL)
}
\arguments{
\item{object}{A `Rational` object}

\item{digits}{The number of digits to print. The returned value has machine
precision.}
}
\value{
The decimal numeric created from the `Rational`
}
\description{
Convert a `Rational` into a decimal
}
\examples{
rat <- R(2, 4)
quotient(rat)
}
