% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rational Class.R
\name{R}
\alias{R}
\title{Rational number creator}
\usage{
R(num, denom)
}
\arguments{
\item{num}{Numerator integer}

\item{denom}{Denominator integer}
}
\value{
A `Rational` object
}
\description{
Rational number creator
}
