% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rational Functions.R
\name{+,Rational,Rational-method}
\alias{+,Rational,Rational-method}
\alias{-,Rational,ANY-method}
\alias{-,Rational,Rational-method}
\alias{*,Rational,Rational-method}
\alias{/,Rational,Rational-method}
\title{Operators for `Rational` Objects}
\usage{
\S4method{+}{Rational,Rational}(e1, e2)

\S4method{-}{Rational,ANY}(e1)

\S4method{-}{Rational,Rational}(e1, e2)

\S4method{*}{Rational,Rational}(e1, e2)

\S4method{/}{Rational,Rational}(e1, e2)
}
\arguments{
\item{e1}{a `Rational` Object}

\item{e2}{a `Rational` Object}
}
\value{
a `Rational` Object
}
\description{
Operators for `Rational` Objects
}
