% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rational Functions.R
\name{show,Rational-method}
\alias{show,Rational-method}
\title{Show method for `Rational` object}
\usage{
\S4method{show}{Rational}(object)
}
\arguments{
\item{object}{A `Rational` object}
}
\value{
`object`, invisibly
}
\description{
Show method for `Rational` object
}
