% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RationalFunctions.R
\name{simplify}
\alias{simplify}
\title{Simplify a `Rational` object}
\usage{
simplify(object)
}
\arguments{
\item{object}{A `Rational` object}
}
\value{
A `Rational` object equal to `object`, but whose numerator and
  denominator have GCD of 1.
}
\description{
Simplify a `Rational` object
}
